+++ node/src/main/kotlin/net/corda/node/services/transactions/ValidatingNotaryFlow.kt
@@ -30,6 +30,7 @@ class ValidatingNotaryFlow(otherSideSession: FlowSession, service: TrustedAuthor
             validateRequestSignature(NotarisationRequest(stx.inputs, stx.id), requestPayload.requestSignature)
             val notary = stx.notary
             checkNotary(notary)
+            checkInputs(stx.inputs)
             resolveAndContractVerify(stx)
             verifySignatures(stx)
             val timeWindow: TimeWindow? = if (stx.coreTransaction is WireTransaction) stx.tx.timeWindow else null
